Hide side-menu content from keyboard navigation
When `.side-menu__content` is closed it's still accessible through keyboard navigation. 
So when tabbing through the header I would expect keyboard focus to move from the side-menu icon to the logo, instead it moves to the links in the `.side-menu__content` element which is positioned outside of the viewport, but not properly hidden. Confusing the user in the process.